requirements.txt
torch
transformers
accelerate
gtts
python-dotenv
protobuf
sentencepiece
numpy
